chore: Remove tsup
All our dependencies are 202.1kB minified and 46.3kB gzipped whilst @typed/fp
including the portions of these dependencies that it includes is 201kb and is 50.3kb
gzipped. I'm pretty happy about that :)

Combined with islands architecture this will lead to very fast loading applications. Especially
when using a CDN for external resources