fix: don't replicate values into descendant Refs
When using the Context API, it is sometimes desirable to ensure the descendant Refs environment
is notified of updates from an Ancestor Refs environment where Context has determined it to originate from.
Utilizing Context.use one can subscribe to the updates of the underlying Ref.

This change ensures that when those updates occur, the Ancestor value will not be stored in the descendant
as this would then break the usage of Context, as it'd now exist in the local environment creating duplicates.