refactor: remove all need of Effect.scoped
This is important when an Fx explicitly requires the use of Scope, we don't want internal operators to close these resources and instead, leave it to the user to decide when these resources should be released.

This is important for @typed/html, as it uses Scope to track event listeners. Even when the Fx ends, there's a decent chance you still want those event listeners to continue to exist.