WIP: rewriting of RefSubject + Computed
Attempts to remove giant amount of duplication in the
RefSubject module to improve maintenance.

This may have a negative impact on performance since it
removes the usage of classes to eliminate closures being constructed.
It may be worth revisiting a class-based approach in the future once
this proves effective.

This PR creates, a mostly internal, FxTransform which sorta works like
a Yoneda but more specialized towards usage with a
Computed/RefSubject value and handling the transformations of
the implemented Fx and .get Effect separately. Maybe its useful
in user-land, but it's got 12 type parameters to contend with.

This RefTransform type is utilized to generalized the capabilities of
Computed and the new data type Filtered. Computed had more
recently been converted to allow `filterMap` which changed its
`.get` to return `Cause.NoSuchElementException`. This is kinda
unwanted in many cases where you're not doing any filtering.

The abilitity to filterMap now returns a Filtered<R, E, A>, which is
just like a Computed, but includes Cause.NoSuchElementExpection in
its `.get` to retrieve its current value.